refactor: Change itemId type from uint16 to uint256 in InventoryLib
- Update GridItem struct to use uint256 itemId for consistency with registries
- Update placeItem() and placeItemWithRotation() function parameters to uint256
- Update InventoryAction struct itemId from uint16 to uint256 in IGameState
- Update _getItemShape() helper function to accept uint256 itemId parameter
- Remove uint16 casting in _placeFishInInventory() function calls
- Replace all placeholder 1x1 item shapes with proper registry shape lookups
- Move hasEquippedFishingRod function from GameStateBase to FishingManager
- All 61 tests passing with improved type consistency

+++ contracts/src/interfaces/IGameState.sol
@@ -34,7 +34,7 @@ struct InventoryAction {
     uint8 toX; // Target position
     uint8 toY;
     uint8 rotation; // 0=up, 1=right, 2=down, 3=left
-    uint16 itemId; // Item being manipulated
+    uint256 itemId; // Item being manipulated
 }
 
 interface IGameState {
